# [SPARK-35003][SQL] Improve performance for reading smallint in vectorized Parquet reader