--- a/sql/core/src/main/java/org/apache/spark/sql/execution/datasources/parquet/VectorizedValuesReader.java
+++ b/sql/core/src/main/java/org/apache/spark/sql/execution/datasources/parquet/VectorizedValuesReader.java
@@ -28,6 +28,7 @@ import org.apache.parquet.io.api.Binary;
 public interface VectorizedValuesReader {
   boolean readBoolean();
   byte readByte();
+  short readShort();
   int readInteger();
   long readLong();
   float readFloat();
@@ -39,6 +40,7 @@ public interface VectorizedValuesReader {
    */
   void readBooleans(int total, WritableColumnVector c, int rowId);
   void readBytes(int total, WritableColumnVector c, int rowId);
+  void readShorts(int total, WritableColumnVector c, int rowId);
   void readIntegers(int total, WritableColumnVector c, int rowId);
   void readIntegersWithRebase(int total, WritableColumnVector c, int rowId, boolean failIfRebase);
   void readUnsignedIntegers(int total, WritableColumnVector c, int rowId);